Vercel/requirements.txt
Flask
pinecone_text.sparse
langchain_community
pinecone
langchain_openai
vercel-flask